Dashboard fixes
In order to make upstream dashboard consumers (grafana) aware of dashboard upgrades, transmit a digest of the dashboard source contents as an attribute of the payload.

Also: guard the dashboard registration with a flag to stop spurious registration.

Fix dashboard source, and disable tagged variables as that has proven brittle